Began Adding in changes to  portfolio page.
For the most part this was mostly just figuring stuff out, like layouts etc. Still not sure what pattern to use for the carousel, need to do some proper research rather than just fumbling around.

My current plan looks something like a linked list, but transitioning between banners wouldn't be as possible. Could make it so that each one knows where it is in the order, and gets a negative or positive value based on whether they're going in or out of view